// src/types/cricket.ts
export interface Scorecard {
  batting: BattingEntry[];
  bowling: BowlingEntry[];
  extras: string;
  totalRuns: number;
  totalWickets: number;
  totalOvers: number;
  inning: string;
}

export interface BattingEntry {
  batsman: BatsmanInfo;
  dismissal: string;
  r: number;
  b: number;
  "4s": number;
  "6s": number;
  sr: string;
}

export interface BatsmanInfo {
  id: string;
  name: string;
}

export interface BowlingEntry {
  bowler: BowlerInfo;
  o: number;
  m: number;
  r: number;
  w: number;
  eco: string;
}

export interface BowlerInfo {
  id: string;
  name: string;
}